2-06 도메인별로 분류하기
